dynamic linker: fix the old bug in the logic to calculate the ELF base
In order to properly calculate and then map an elf object
in virtual memory, the set_base() method needs to identify the lowest
and the highest PT_LOAD header in terms of the p_vaddr value.

Unfortunately, there seems to be a bug in how it is implemented now
and in some cases, when for example there are multiple PT_LOAD and non-PT_LOAD
headers sharing the same p_vaddr value, it ends up using wrong
header and then overwriting other segments data.

This patch fixes it by separating filtering of PT_LOAD headers
and indentifying the lowest and highest ones.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20200708184200.33634-1-jwkozaczuk@gmail.com>